move more control functionality into details
